Let FFI implementation files share their types
